==> 01_postgresql/04_json_read.ipynb <==
#Error simulation
import pandas as pd
pd.set_option('display.max_columns', None)
caminho_do_arquivo = r"D:\engenharia de dados\curso udemy\Arquivos\01.+Postgree\01. Postgree\Origem de dados"
df = pd.read_json(caminho_do_arquivo)
df.head

in Brazilian Portuguese you need to enter utf-8 enconding, the error was caused by special characters

import pandas as pd
caminho_do_arquivo = r"D:\engenharia de dados\curso udemy\Arquivos\01.+Postgree\01. Postgree\Origem de dados\V_OCORRENCIA_AMPLA.json"
df = pd.read_json(caminho_do_arquivo, encoding='utf-8-sig')
df.head()

import codecs
from chardet.universaldetector import UniversalDetector

caminho_do_arquivo = r"D:\engenharia de dados\curso udemy\Arquivos\01.+Postgree\01. Postgree\Origem de dados\V_OCORRENCIA_AMPLA.json"
detector = UniversalDetector()

with codecs.open(caminho_do_arquivo, 'rb') as arquivo:
    for linha in arquivo:
        detector.feed(linha)
        if detector.done:
            break

detector.close()
encoding_detectado = detector.result['encoding']

print(encoding_detectado)

import json
import codecs

caminho_do_arquivo = r"D:\engenharia de dados\curso udemy\Arquivos\01.+Postgree\01. Postgree\Origem de dados\V_OCORRENCIA_AMPLA.json"

with codecs.open(caminho_do_arquivo, 'r', encoding='utf-8-sig') as arquivo:
    data = json.load(arquivo)
df = pd.json_normalize(data)
df.head()

Loading automatically encoding

pd.set_option('display.max_columns', None)
caminho_do_arquivo = r"D:\engenharia de dados\curso udemy\Arquivos\01.+Postgree\01. Postgree\Origem de dados\V_OCORRENCIA_AMPLA.json"
detector = UniversalDetector()

with codecs.open(caminho_do_arquivo, 'rb') as arquivo:
    for linha in arquivo:
        detector.feed(linha)
        if detector.done:
            break
detector.close()
encoding_detectado = detector.result['encoding']

df_json_automatico = pd.read_json(caminho_do_arquivo, encoding=encoding_detectado)
df.head()
