==> 01_postgresql/05_project_anac.ipynb <==
Importing DataFrame

import pandas as pd
pd.set_option('display.max_columns', None)
caminho_do_arquivo = r"D:\engenharia de dados\curso udemy\Arquivos\01.+Postgree\01. Postgree\Origem de dados\V_OCORRENCIA_AMPLA.json"
df = pd.read_json(caminho_do_arquivo, encoding='utf-8-sig')
df.head(3)

Filtering and modifying columns

df.columns

colunas = ['Numero_da_Ocorrencia', 'Classificacao_da_Ocorrência', 'Data_da_Ocorrencia', 'Municipio', 'UF', 'Regiao', 'Nome_do_Fabricante'] 
df = df[colunas]

df.rename(columns={ 'Classificacao_da_Ocorrência' : 'Classificacao_da_Ocorrencia' }, inplace=True)

df.head(3)

Inserting data from the dataframe into postgre

import psycopg2

dbname = 'python'
user = 'postgres'
password = 'teste123'
host = 'localhost'
port = '5432'

conn = psycopg2.connect(dbname=dbname, user=user, password=password, host=host, port=port)

#Delete base before insert
cur = conn.cursor()

cur.execute("delete from anac ")

for indice,linha in df.iterrows():
    cur.execute("""
                insert into Anac (     
                    Numero_da_Ocorrencia, 
                    Classificacao_da_Ocorrencia, 
                    Data_da_Ocorrencia, 
                    Municipio, 
                    UF, 
                    Regiao, 
                    Nome_do_Fabricante
                ) VALUES (%s, %s, %s, %s, %s, %s, %s)
                
                """,(
                    linha["Numero_da_Ocorrencia"],
                    linha["Classificacao_da_Ocorrencia"],
                    linha["Data_da_Ocorrencia"],
                    linha["Municipio"],
                    linha["UF"],
                    linha["Regiao"],
                    linha["Nome_do_Fabricante"]
                )
                
                )

conn.commit()

cur.close()
conn.close()
