==> 01_postgresql/08_sqlachemy_sending.ipynb <==
import pandas as pd
pd.set_option('display.max_columns', None)
caminho_do_arquivo = r"D:\engenharia de dados\curso udemy\Arquivos\01.+Postgree\01. Postgree\Origem de dados\V_OCORRENCIA_AMPLA.json"
df = pd.read_json(caminho_do_arquivo, encoding='utf-8-sig')
df.head(3)

colunas = ["Numero_da_Ocorrencia", "Classificacao_da_Ocorrência", "Data_da_Ocorrencia","Municipio","UF","Regiao","Nome_do_Fabricante","Modelo"]
df = df[colunas]
df.rename( columns={  'Classificacao_da_Ocorrência' : 'Classificacao_da_Ocorrencia'  } ,inplace=True )
df.head()


View data types

df.dtypes

Connecting to postgre using sqlalchemy


from sqlalchemy import create_engine ,Integer, String, Date,VARCHAR

dbname   = 'python'
user     = 'postgres'
password = 'teste123'
host     = 'localhost'
port     = '5432' 

conn = f'postgresql://{user}:{password}@{host}:{port}/{dbname}'

engine = create_engine(conn)

nome_tabela = 'anac_sqlalchemy'

df.to_sql(nome_tabela, engine, index=False, if_exists='replace', dtype={ 
    'Numero_da_Ocorrencia' : Integer,
    'Classificacao_da_Ocorrencia' : VARCHAR(50),
    'Data_da_Ocorrencia' : Date,
    'Municipio' : VARCHAR(100),
    'UF' : VARCHAR(50),
    'Regiao' : VARCHAR(50),
    'Nome_do_Fabricante' : VARCHAR(100),
    'Modelo' : VARCHAR(50)    
})

engine.dispose()
