==> 01_postgresql/13_saving_different_formats.ipynb <==
Saving the file in different formats

from sqlalchemy import create_engine
import pandas as pd
#conexao com banco de dados
dbname   = 'python'
user     = 'postgres'
password = 'teste123'
host     = 'localhost'
port     = '5432' 

conexao_str = f'postgresql://{user}:{password}@{host}:{port}/{dbname}'
engine = create_engine(conexao_str)


query = """
select 
id
,dt_ocorrencia
,uf
,regiao
,classificacao_da_ocorrencia
from public.anac_mapeamento


"""
df=pd.read_sql_query(query,engine )
df.head(5)


JSON

pasta_destino_json = r"D:\engenharia de dados\curso udemy\Github\python_postgresql\destiny\Anac.json"
df.to_json(pasta_destino_json, orient="records", lines=True)

Excel

pasta_destino_excel = r"D:\engenharia de dados\curso udemy\Github\python_postgresql\destiny\Anac.xlsx"
df.to_excel(pasta_destino_excel, index = False, sheet_name="Base")

CSV

pasta_destino_csv = r"D:\engenharia de dados\curso udemy\Github\python_postgresql\destiny\Anac.csv"
df.to_csv(pasta_destino_csv,sep=',',index=False)

Parquet

#!pip install pyarrow
#!pip install fastparquet

pasta_destino_parquet = r"D:\engenharia de dados\curso udemy\Github\python_postgresql\destiny\Anac.parquet"
df.to_parquet(pasta_destino_parquet, index=False, engine='pyarrow')

df.head()

df.columns

Parquet partitioned

pasta_destino_particionado = r"D:\engenharia de dados\curso udemy\Github\python_postgresql\destiny\Anac_partitioned.parquet"
colunas_particionamento = ['regiao','uf']

df.to_parquet(pasta_destino_particionado, index=False,
              partition_cols=colunas_particionamento,
              engine='pyarrow'
              )
